+++ src/Api/Providers/Pinners.php
@@ -37,7 +37,9 @@ class Pinners extends Provider
      */
     public function info($username)
     {
-        return $this->execGetRequest(['username' => $username], UrlHelper::RESOURCE_USER_INFO);
+        return $this->execGetRequest(
+            ['username' => $username], UrlHelper::RESOURCE_USER_INFO
+        );
     }
 
     /**