+++ src/Api/Traits/HandlesRequestAndResponse.php
@@ -5,6 +5,10 @@ namespace seregazhuk\PinterestBot\Api\Traits;
 use seregazhuk\PinterestBot\Api\Request;
 use seregazhuk\PinterestBot\Api\Response;
 
+/**
+ * Trait HandlesRequestAndResponse
+ * @package seregazhuk\PinterestBot\Api\Traits
+ */
 trait HandlesRequestAndResponse
 {
     /**